Prevent loading plugin in headless context
